refactor: separate gridded api service into new file
